+++ client/upgrade/cli/query.go
@@ -2,12 +2,16 @@ package cli
 
 import (
 	"fmt"
+	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/wire"
 	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
 	"github.com/irisnet/irishub/client/context"
+	upgcli "github.com/irisnet/irishub/client/upgrade"
 	"github.com/irisnet/irishub/modules/upgrade"
 	"github.com/irisnet/irishub/modules/upgrade/params"
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"os"
 )
 
@@ -22,8 +26,8 @@ func GetInfoCmd(storeName string, cdc *wire.Codec) *cobra.Command {
 				WithLogger(os.Stdout).
 				WithAccountDecoder(authcmd.GetAccountDecoder(cdc))
 
-			res_height, _ := cliCtx.QueryStore([]byte("gov/"+upgradeparams.ProposalAcceptHeightParameter.GetStoreKey()), "params")
-			res_proposalID, _ := cliCtx.QueryStore([]byte("gov/"+upgradeparams.CurrentUpgradeProposalIdParameter.GetStoreKey()), "params")
+			res_height, _ := cliCtx.QueryStore([]byte(upgradeparams.ProposalAcceptHeightParameter.GetStoreKey()), "params")
+			res_proposalID, _ := cliCtx.QueryStore([]byte(upgradeparams.CurrentUpgradeProposalIdParameter.GetStoreKey()), "params")
 			var height int64
 			var proposalID int64
 			cdc.MustUnmarshalBinary(res_height, &height)
@@ -36,16 +40,59 @@ func GetInfoCmd(storeName string, cdc *wire.Codec) *cobra.Command {
 			res_version, _ := cliCtx.QueryStore(upgrade.GetVersionIDKey(versionID), storeName)
 			var version upgrade.Version
 			cdc.MustUnmarshalBinary(res_version, &version)
-			output, err := wire.MarshalJSONIndent(cdc, version)
+
+			upgradeInfoOutput := upgcli.ConvertDepositToDepositOutput(version, proposalID, height)
+
+			output, err := wire.MarshalJSONIndent(cdc, upgradeInfoOutput)
 			if err != nil {
 				return err
 			}
 
 			fmt.Println(string(output))
-			fmt.Println("CurrentProposalId           = ", proposalID)
-			fmt.Println("CurrentProposalAcceptHeight = ", height)
 			return nil
 		},
 	}
+	return cmd
+}
+
+// Command to Get a Switch Information
+func GetCmdQuerySwitch(storeName string, cdc *wire.Codec) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "query-switch",
+		Short: "query switch details",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			proposalID := viper.GetInt64(flagProposalID)
+			voterStr := viper.GetString(flagVoter)
+
+			voter, err := sdk.AccAddressFromBech32(voterStr)
+			if err != nil {
+				return err
+			}
+
+			cliCtx := context.NewCLIContext().
+				WithCodec(cdc).
+				WithLogger(os.Stdout).
+				WithAccountDecoder(authcmd.GetAccountDecoder(cdc))
+
+			res, err := cliCtx.QueryStore(upgrade.GetSwitchKey(proposalID, voter), storeName)
+			if len(res) == 0 || err != nil {
+				return errors.Errorf("proposalID [%d] is not existed", proposalID)
+			}
+
+			var switchMsg upgrade.MsgSwitch
+			cdc.MustUnmarshalBinary(res, &switchMsg)
+			output, err := wire.MarshalJSONIndent(cdc, switchMsg)
+			if err != nil {
+				return err
+			}
+
+			fmt.Println(string(output))
+			return nil
+		},
+	}
+
+	cmd.Flags().String(flagProposalID, "", "proposalID of upgrade swtich being queried")
+	cmd.Flags().String(flagVoter, "", "Address sign the switch msg")
+
 	return cmd
 }