+++ modules/gov/genesis.go
@@ -3,18 +3,20 @@ package gov
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/irisnet/irishub/modules/gov/params"
-	"github.com/irisnet/irishub/modules/parameter"
+	"github.com/irisnet/irishub/modules/iparam"
+	"fmt"
+	"github.com/irisnet/irishub/types"
 )
 
-// GenesisState - all staking state that must be provided at genesis
+// GenesisState - all gov state that must be provided at genesis
 type GenesisState struct {
-	StartingProposalID int64                      `json:"starting_proposalID"`
-	DepositProcedure   govparams.DepositProcedure `json:"deposit_period"`
-	VotingProcedure    govparams.VotingProcedure            `json:"voting_period"`
-	TallyingProcedure  TallyingProcedure          `json:"tallying_procedure"`
+	StartingProposalID int64                        `json:"starting_proposalID"`
+	DepositProcedure   govparams.DepositProcedure   `json:"deposit_period"`
+	VotingProcedure    govparams.VotingProcedure    `json:"voting_period"`
+	TallyingProcedure  govparams.TallyingProcedure  `json:"tallying_procedure"`
 }
 
-func NewGenesisState(startingProposalID int64, dp govparams.DepositProcedure, vp govparams.VotingProcedure, tp TallyingProcedure) GenesisState {
+func NewGenesisState(startingProposalID int64, dp govparams.DepositProcedure, vp govparams.VotingProcedure, tp govparams.TallyingProcedure) GenesisState {
 	return GenesisState{
 		StartingProposalID: startingProposalID,
 		DepositProcedure:   dp,
@@ -31,23 +33,73 @@ func InitGenesis(ctx sdk.Context, k Keeper, data GenesisState) {
 		panic(err)
 	}
 	//k.setDepositProcedure(ctx, data.DepositProcedure)
-	parameter.InitGenesisParameter(&govparams.DepositProcedureParameter, ctx, data.DepositProcedure)
-	parameter.InitGenesisParameter(&govparams.VotingProcedureParameter, ctx, data.VotingProcedure)
-	k.setTallyingProcedure(ctx, data.TallyingProcedure)
+	iparam.InitGenesisParameter(&govparams.DepositProcedureParameter, ctx, data.DepositProcedure)
+	iparam.InitGenesisParameter(&govparams.VotingProcedureParameter, ctx, data.VotingProcedure)
+	iparam.InitGenesisParameter(&govparams.TallyingProcedureParameter, ctx, data.TallyingProcedure)
 
 }
 
 // WriteGenesis - output genesis parameters
 func WriteGenesis(ctx sdk.Context, k Keeper) GenesisState {
 	startingProposalID, _ := k.getNewProposalID(ctx)
-	depositProcedure := k.GetDepositProcedure(ctx)
-	votingProcedure := k.GetVotingProcedure(ctx)
-	tallyingProcedure := k.GetTallyingProcedure(ctx)
+	depositProcedure := govparams.GetDepositProcedure(ctx)
+	votingProcedure := govparams.GetVotingProcedure(ctx)
+	tallyingProcedure := govparams.GetTallyingProcedure(ctx)
 
 	return GenesisState{
 		StartingProposalID: startingProposalID,
 		DepositProcedure:   depositProcedure,
 		VotingProcedure:    votingProcedure,
 		TallyingProcedure:  tallyingProcedure,
+	}
+}
+
+// get raw genesis raw message for testing
+func DefaultGenesisState() GenesisState {
+	Denom  := "iris"
+	IrisCt := types.NewDefaultCoinType(Denom)
+	minDeposit, err := IrisCt.ConvertToMinCoin(fmt.Sprintf("%d%s", 1000, Denom))
+	if err != nil {
+		panic(err)
+	}
+	return GenesisState{
+		StartingProposalID: 1,
+		DepositProcedure: govparams.DepositProcedure{
+			MinDeposit:       sdk.Coins{minDeposit},
+			MaxDepositPeriod: 20000,
+		},
+		VotingProcedure: govparams.VotingProcedure{
+			VotingPeriod: 20000,
+		},
+		TallyingProcedure: govparams.TallyingProcedure{
+			Threshold:         sdk.NewRat(1, 2),
+			Veto:              sdk.NewRat(1, 3),
+			GovernancePenalty: sdk.NewRat(1, 100),
+		},
+	}
+}
+
+// get raw genesis raw message for testing
+func DefaultGenesisStateForTest() GenesisState {
+	Denom  := "iris"
+	IrisCt := types.NewDefaultCoinType(Denom)
+	minDeposit, err := IrisCt.ConvertToMinCoin(fmt.Sprintf("%d%s", 10, Denom))
+	if err != nil {
+		panic(err)
+	}
+	return GenesisState{
+		StartingProposalID: 1,
+		DepositProcedure: govparams.DepositProcedure{
+			MinDeposit:       sdk.Coins{minDeposit},
+			MaxDepositPeriod: 10,
+		},
+		VotingProcedure: govparams.VotingProcedure{
+			VotingPeriod: 10,
+		},
+		TallyingProcedure: govparams.TallyingProcedure{
+			Threshold:         sdk.NewRat(1, 2),
+			Veto:              sdk.NewRat(1, 3),
+			GovernancePenalty: sdk.NewRat(1, 100),
+		},
 	}
 }